+++ R/dlg_save.R
@@ -137,10 +137,11 @@ gui = .GUI) {
   invisible(gui)
 }
 
-#' @inheritParams get_syst
+#' @inheritParams get_system
 #' @export
 #' @rdname dlg_save
-dlgSave.nativeGUI <- function(default, title, filters = dlg_filters["All", ], rstudio = TRUE, ..., gui = .GUI) {
+dlgSave.nativeGUI <- function(default, title, filters = dlg_filters["All", ],
+rstudio = TRUE, ..., gui = .GUI) {
   # The native version of the file save dialog box
   gui$setUI(widgets = "nativeGUI")
   # A 'save file' dialog box
@@ -149,8 +150,7 @@ dlgSave.nativeGUI <- function(default, title, filters = dlg_filters["All", ], rs
   #
   # It is a replacement for choose.files(), tkgetSaveFile()
   # & file.choose(new = TRUE), not implemented yet in R 2.14, by the way
-  syst <- get_syst(rstudio)
-  res <- switch(syst,
+  res <- switch(get_system(rstudio),
     RStudio = .rstudio_dlg_save(gui$args$default, gui$args$title,
       gui$args$filters),
     Windows = .win_dlg_save(gui$args$default, gui$args$title, gui$args$filters),